@@ -1,34 +1,26 @@
 {
-  "name": "template-svelte",
-  "version": "1.0.0",
-  "description": "A Svelte app template for bootstrapping deployments on Orb's HyperEdge",
-  "repository": "https://github.com/orbclouds/template-svelte.git",
-  "author": "Orb <orb@orbclouds.com>",
-  "license": "MIT",
-  "private": false,
+  "name": "frontend",
+  "version": "0.0.1",
   "scripts": {
-    "serve": "serve -s build",
-    "build": "snowpack build && rm -rf build/dist build/_snowpack",
-    "start": "snowpack dev"
+    "dev": "svelte-kit dev",
+    "build": "svelte-kit build",
+    "package": "svelte-kit package",
+    "preview": "svelte-kit preview",
+    "check": "svelte-check --tsconfig ./tsconfig.json",
+    "check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
+    "lint": "prettier --ignore-path .gitignore --check --plugin-search-dir=. .",
+    "format": "prettier --ignore-path .gitignore --write --plugin-search-dir=. ."
   },
   "devDependencies": {
-    "@beyonk/svelte-google-analytics": "^2.0.0",
-    "@hisystems/snowpack-plugin-postcss": "^1.0.3",
-    "@snowpack/plugin-dotenv": "^2.0.5",
-    "@snowpack/plugin-run-script": "^2.3.0",
-    "@snowpack/plugin-svelte": "^3.5.2",
-    "@snowpack/plugin-typescript": "^1.1.0",
-    "@snowpack/plugin-webpack": "^2.3.0",
-    "@tsconfig/svelte": "^1.0.10",
-    "@types/snowpack-env": "^2.3.3",
-    "autoprefixer": "^10.0.4",
-    "onchange": "^7.1.0",
-    "postcss": "^8.1.14",
-    "postcss-cli": "^8.3.0",
-    "serve": "^11.3.2",
-    "snowpack": "^3.0.11",
-    "svelte": "^3.31.2",
-    "svelte-preprocess": "^4.6.3",
-    "typescript": "^4.0.5"
-  }
+    "@sveltejs/adapter-auto": "next",
+    "@sveltejs/kit": "next",
+    "prettier": "^2.4.1",
+    "prettier-plugin-svelte": "^2.4.0",
+    "svelte": "^3.44.0",
+    "svelte-check": "^2.2.6",
+    "svelte-preprocess": "^4.9.4",
+    "tslib": "^2.3.1",
+    "typescript": "^4.4.3"
+  },
+  "type": "module"
 }
