@@ -1,12 +1,18 @@
-const preprocess = require('svelte-preprocess');
+import adapter from '@sveltejs/adapter-auto';
+import preprocess from 'svelte-preprocess';
 
-module.exports = {
-  preprocess: preprocess({
-    defaults: {
-      script: 'typescript',
-    },
-  }),
-  compilerOptions: {
-    css: false,
-  },
+/** @type {import('@sveltejs/kit').Config} */
+const config = {
+	// Consult https://github.com/sveltejs/svelte-preprocess
+	// for more information about preprocessors
+	preprocess: preprocess(),
+
+	kit: {
+		adapter: adapter(),
+
+		// hydrate the <div id="svelte"> element in src/app.html
+		target: '#svelte'
+	}
 };
+
+export default config;
